% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derived.traits.2014.R
\name{igg.uplc.derived.traits.2014}
\alias{igg.uplc.derived.traits.2014}
\title{Derived traits for Glycan peaks in IgG for UPLC
based on paper from 2014.}
\usage{
igg.uplc.derived.traits.2014(data)
}
\arguments{
\item{data}{data frame that holds columns representing Glycans.
These column names should start with 'GP'.}
}
\value{
Returns a data.frame with derived traits
}
\description{
Calcuates values of derived traits for Glycan peaks in IgG for UPLC
}
\author{
Ivo Ugrina; Frano Vučković
}
\references{
Jennifer E. Huffman et al.
"Comparative Performance of Four Methods for High-throughput Glycosylation Analysis of Immunoglobulin G in Genetic and Epidemiological Research*"
\link{http://dx.doi.org/10.1074/mcp.M113.037465 }
}

