% Generated by roxygen2 (4.0.2): do not edit by hand
\name{igg.uplc.derived.traits}
\alias{igg.uplc.derived.traits}
\title{Derived traits for Glycan peaks in IgG for UPLC}
\usage{
igg.uplc.derived.traits(data)
}
\arguments{
\item{data}{data frame which holds columns representing Glycans.
These column names should start with 'GP'.}
}
\value{
Returns a data.frame with derived traits
}
\description{
Calcuates values of derived traits for Glycan peaks in IgG for UPLC
}
\author{
Ivo Ugrina; Frano Vučković
}
\references{
Jennifer E. Huffman et al.
"Comparative Performance of Four Methods for High-throughput Glycosylation Analysis of Immunoglobulin G in Genetic and Epidemiological Research*"
\link{http://dx.doi.org/10.1074/mcp.M113.037465 }
}

