% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ildt.translate}
\alias{ildt.translate}
\title{Translate names between computer readable and human readable
for derived traits of IgG with LCMS}
\usage{
ildt.translate(orignames, method = "2014")
}
\arguments{
\item{orignames}{vector; type string}
}
\value{
Returns a character vector with original and translated names
}
\description{
Translates names between computer readable and human readable
for derived traits of IgG with LCMS
}
\author{
Ivo Ugrina
}

