% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iudt.translate.2014}
\alias{iudt.translate.2014}
\title{Translate names between computer readable and human readable
for derived traits of IgG with UPLC
based on an article from 2014.}
\usage{
iudt.translate.2014(orignames)
}
\arguments{
\item{orignames}{vector; type string}
}
\value{
Returns a character vector with original and translated names
}
\description{
Translates names between computer readable and human readable
for derived traits of IgG with UPLC
}
\author{
Ivo Ugrina
}
\references{
Jennifer E. Huffman et al.
"Comparative Performance of Four Methods for High-throughput Glycosylation Analysis of Immunoglobulin G in Genetic and Epidemiological Research*"
\link{http://dx.doi.org/10.1074/mcp.M113.037465 }
}

