% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glycanrinfo.R
\docType{data}
\name{mpiu}
\alias{mpiu}
\title{Multiple plates IgG UPLC data example}
\format{\preformatted{'data.frame':	13680 obs. of  4 variables:
 $ Plate : int  1 1 1 1 1 1 1 1 1 1 ...
 $ gid   : chr  "1_1" "1_1" "1_1" "1_1" ...
 $ glycan: Factor w/ 24 levels "GP1","GP2","GP3",..: 1 10 11 12 13 14 15 16 17 18 ...
 $ value : num  35277 1657042 269981 211240 87462 ...
}}
\usage{
mpiu
}
\description{
This is an example of the data obtained by UPLC
while analysing IgG.
}
\keyword{datasets}

