% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glycanrinfo.R
\docType{data}
\name{mpiunorm}
\alias{mpiunorm}
\title{Normalizied mpiu data}
\format{\preformatted{'data.frame':	68400 obs. of  5 variables:
 $ Plate : int  1 1 1 1 1 1 1 1 1 1 ...
 $ gid   : chr  "1_1" "1_1" "1_1" "1_1" ...
 $ norm  : chr  "mediannorm" "mediannorm" "mediannorm" "mediannorm" ...
 $ glycan: Factor w/ 24 levels "GP1","GP2","GP3",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ value : num  -0.262 -0.167 -0.247 6.744 -0.247 ...
}}
\usage{
mpiunorm
}
\description{
This is an example of the normalized mpiu data.
It is intended to be used for unit tests.
}
\keyword{datasets}

