% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derived.traits.R
\name{phdt}
\alias{phdt}
\title{Derived traits for Glycan peaks in PLASMA for HPLC}
\usage{
phdt(data, method = "2011")
}
\arguments{
\item{data}{data frame that holds columns representing Glycans.
These column names should start with 'GP'.}

\item{method}{year of the derived traits definition. By default 2011.}
}
\value{
Returns a data.frame with derived traits
}
\description{
Calcuates values of derived traits for Glycan peaks in Plasma for HPLC
}
\details{
Calculates derived traits from basic glycan peaks. User can choose
which definition of the derived traits he will use
(see references for different versions/definitions of derived traits).
}
\author{
Ivo Ugrina
}
\references{
Lu et al. (2011)
"Screening Novel Biomarkers for Metabolic Syndrome by Profiling
 Human Plasma N-Glycans in Chinese Han and Croatian Populations"
\url{http://dx.doi.org/10.1021/pr2004067}
}

