% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plasma.hplc.derived.traits}
\alias{plasma.hplc.derived.traits}
\title{Derived traits for Glycan peaks in PLASMA for HPLC}
\usage{
plasma.hplc.derived.traits(d, method = "2011")
}
\arguments{
\item{data}{data frame that holds columns representing Glycans.
These column names should start with 'GP'.}
}
\value{
Returns a data.frame with derived traits
}
\description{
Calcuates values of derived traits for Glycan peaks in Plasma for HPLC
}
\author{
Ivo Ugrina
}

