% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{quantilenorm}
\alias{quantilenorm}
\title{Quantile Normalization of glycan data}
\usage{
quantilenorm(d, subclasses = FALSE, transpose = FALSE)
}
\arguments{
\item{d}{data frame in long format containing glycan measurements}

\item{subclasses}{should data be normalized per subclass}

\item{transpose}{transpose the data prior to normalization}
}
\value{
Returns a data.frame with original glycan values substituted by normalized ones
}
\description{
Returns glycans normalized with Quantile Normalization approach.
}
\details{
Input data frame should have at least the following three columns:
  - gid - representing a unique name of a sample
  - glycan - representing glycan names
  - value - representing measured values
and if the subclasses argument is given it should also have column:
  - isoform - representing subclasses (e.g. IgG1, IgG2 and IgG4)
}
\author{
Ivo Ugrina, Lucija Klarić
}

