% Generated by roxygen2 (4.0.2): do not edit by hand
\name{igg.lcms.derived.traits}
\alias{igg.lcms.derived.traits}
\title{Derived traits for Glycan peaks in IgG for LCMS}
\usage{
igg.lcms.derived.traits(data, method = "2014")
}
\arguments{
\item{data}{data frame that holds columns representing Glycans.}
}
\value{
Returns a data.frame with derived traits
}
\description{
Calcuates values of derived traits for Glycan peaks in IgG for LCMS
}
\author{
Ivo Ugrina
}

