% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derived.traits.R
\name{igg.uplc.derived.traits}
\alias{igg.uplc.derived.traits}
\title{Derived traits for Glycan peaks in IgG for UPLC}
\usage{
igg.uplc.derived.traits(data, method = "2014")
}
\arguments{
\item{data}{data frame that holds columns representing Glycans.
These column names should start with 'GP'.}
}
\value{
Returns a data.frame with derived traits
}
\description{
Calcuates values of derived traits for Glycan peaks in IgG for UPLC
}
\author{
Ivo Ugrina
}

