% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycanrinfo.R
\docType{data}
\name{mpiu}
\alias{mpiu}
\title{Multiple plates IgG UPLC data example}
\format{
An object of class \code{data.frame} with 13680 rows and 4 columns.
}
\usage{
mpiu
}
\description{
This is an example of the data obtained by UPLC
while analysing IgG.
}
\keyword{datasets}
