% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycanrinfo.R
\docType{data}
\name{mpiunorm}
\alias{mpiunorm}
\title{Normalized mpiu data}
\format{
An object of class \code{data.frame} with 68400 rows and 5 columns.
}
\usage{
mpiunorm
}
\description{
This is an example of the normalized mpiu data.
It is intended to be used for unit tests.
}
\keyword{datasets}
