% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plasma.hplc.derived.traits.2011}
\alias{plasma.hplc.derived.traits.2011}
\title{Derived traits for Glycan peaks in PLASMA for HPLC
based on paper from 2011.}
\usage{
plasma.hplc.derived.traits.2011(d)
}
\arguments{
\item{data}{data frame which holds columns representing Glycans.
These column names should start with 'GP'.}
}
\value{
Returns a data.frame with derived traits
}
\description{
Calcuates values of derived traits for Glycan peaks in Plasma for HPLC
}
\author{
Ivo Ugrina; Lucija Klarić
}
\references{
Lu et al.
"Screening Novel Biomarkers for Metabolic Syndrome by Profiling
 Human Plasma N-Glycans in Chinese Han and Croatian Populations"
\link{http://dx.doi.org/10.1021/pr2004067}
}

