% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{tanorm}
\alias{tanorm}
\title{Total Area Normalization of glycan data}
\usage{
tanorm(d, subclasses = FALSE)
}
\arguments{
\item{d}{data frame in long format containing glycan measurements}

\item{subclasses}{should data be normalized per subclass}
}
\value{
Returns a data.frame with original glycan values substituted by normalized ones
}
\description{
Returns glycans normalized with Total Area Normalization approach.
}
\details{
Input data frame should have at least the following three columns: \cr
  - gid - representing a unique name of a sample \cr
  - glycan - representing glycan names \cr
  - value - representing measured values \cr
and if the subclasses argument is given it should also have column: \cr
  - isoform - representing subclasses (e.g. IgG1, IgG2 and IgG4)
}
\examples{
data(mpiu)
mpiun <- tanorm(mpiu)
head(mpiun)
}
\author{
Ivo Ugrina
}

